This PR completes the v1.2–v1.4 development cycle.

Major improvements include:

- Refactored the project into a more modular architecture.
- Improved parser infrastructure and internal abstractions.
- Added a complete preprocessing and evaluation pipeline.
- Enhanced configuration, CLI, and project organization.
- Improved documentation and developer workflow.
- Fixed numerous implementation issues and increased overall robustness.

This establishes a stable engineering foundation for future work on grammar induction, category discovery, and semantic parsing.

+++ include/scf/tree_solver.hpp
@@ -11,7 +11,9 @@ namespace scf {
 
 struct SpanScore {
     Span span{};
-    std::uint32_t score{};
+    // Integer objective units: |W| for raw_count, round(strength * 1e12) for
+    // normalized objectives (see kStrengthScale in evidence_builder.hpp).
+    std::uint64_t score{};
 };
 
 struct TreeSolveResult {
